FEATURE : 운동 기록 Dto 추가 및 수정
일별 운동 기록 조회 DTO 개발
DTO 패키지 구조 변경 : recordDto 패키지
