+++ API_container/MadPtApi/src/main/java/com/example/MadPtApi/dto/recordDto/DailyRecordResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.MadPtApi.dto;
+package com.example.MadPtApi.dto.recordDto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -6,22 +6,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RecordDto {
-    @JsonProperty("exercise_id")
-    private Long exerciseId;
+public class DailyRecordResponseDto {
+    @JsonProperty("exercise_name")
+    private String exerciseName;
     @JsonProperty("start_time")
     private Long startTime;
-    @JsonProperty("end_time")
-    private Long endTime;
     private int reps;
     private int sets;
-    // reps를 list로 받아야할듯
-    @JsonProperty("avg_score")
-    private double avgScore;
+    @JsonProperty("burned_kcal")
+    private double burnedKcal;
 }
