+++ API_container/MadPtApi/src/main/java/com/example/MadPtApi/dto/recordDto/RecordSaveRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.MadPtApi.dto;
+package com.example.MadPtApi.dto.recordDto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
